XOR-Decrypt.ps1
XOR-Decrypt.ps1 is currently on my short-list of favorite scripts. Why?
Because it's does some interesting things that I'd not played with
previously. Rather than simply brute-force a XOR encrypted string with a
bunch of single byte and returning results for all of them, it runs each
result through a set of functions that score the result and the scores
are added together and the script picks the one with the highest score
and returns it, as it is most likely the one to be English text. The
functions that make up the score are English letter frequency, English
bigram frequency and English trigram frequency. The script also
calculates entropy, but that does not currently factor into the overall
score, I may change that, though it does well as it is today.
What sucks? It only works for XOR'd strings that have been encrypted by
a single ASCII printable byte value.