XOR-Test.ps1
A script that encrypts one of three text selections with a randomly
generated key up to the length given by MaxKeySize, then it runs the
Hamming Distance calculations against the ciphertext and returns the
data. I've been using this to see if I could come up with improvements
on Normalized Hamming Distance alone.