Created a branch for building some ngram testing
Goal is to add a new flag to enable ngram scoring.
Example: If a XOR operation produces a t, the script should favor
the next key byte that produces and h, since th is the most
frequenly occuring English bigram.
bigrams and trigrams should be supported using something like an
-ngram 2 or -ngram 3 argument. Obvious -ngram 3 will require more
processing and memory than -ngram 2.